services/monitoring/requirements.txt
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
mlflow==2.5.0
evidently==0.4.9
scikit-learn==1.3.0
joblib==1.3.1
cloudpickle==2.2.1
